+++ src/components/NavSidebar.tsx
@@ -1,9 +1,9 @@
 import { FolderOpen, Home, LayoutDashboard, Plus } from 'lucide-react';
 
 // import * as React from 'react';
-import { Button } from '@/components/ui/button';
-import { ScrollArea } from '@/components/ui/scroll-area';
-import { Separator } from '@/components/ui/separator';
+import { Button } from '@/components/ui/Button';
+import { ScrollArea } from '@/components/ui/ScrollArea';
+import { Separator } from '@/components/ui/Separator';
 // import { cn } from '@/lib/utils';
 
 export function NavSidebar() {
